+++ src/screens/SettingsScreen/SettingsStackScreen.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import SettingsMenu from './SettingsMenu';
-import SettingsAudioMenu from './SettingsAudioMenu';
+import SettingsScreen from './SettingsScreen';
+import SettingsAudioScreen from './SettingsAudioScreen';
 
 const Stack = createNativeStackNavigator();
 
 const SettingsStackMenu = () => {
   return (
     <Stack.Navigator>
-      <Stack.Screen name="Main" component={SettingsMenu} options={{ headerShown: false }} />
-      <Stack.Screen name="Audio" component={SettingsAudioMenu} options={{ headerShown: false }} />
+      <Stack.Screen name="Main" component={SettingsScreen} options={{ headerShown: false }} />
+      <Stack.Screen name="Audio" component={SettingsAudioScreen} options={{ headerShown: false }} />
     </Stack.Navigator>
   );
 };
